ClipScript AI/requirements.txt
python-telegram-bot==22.7
yt-dlp==2026.3.17
requests>=2.32.2
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
deepgram-sdk==3.2.0
